+++ example/http2/server/cmd/server.go
@@ -3,25 +3,29 @@ package main
 import (
 	"bytes"
 	"fmt"
+	"net/http"
+	"time"
+)
+
+import (
 	"github.com/dubbogo/triple/pkg/common/logger/default_logger"
 	"github.com/dubbogo/triple/pkg/http2"
 	"github.com/dubbogo/triple/pkg/http2/config"
-	"time"
 )
 
 func main() {
 	svr := http2.NewHttp2Server("localhost:1999", config.ServerConfig{
 		Logger: default_logger.GetDefaultLogger(),
 	})
-	svr.RegisterHandler("/unary", func(path string, header map[string]string, recvChan chan *bytes.Buffer, sendChan chan *bytes.Buffer, ctrlCh chan map[string]string) {
+	svr.RegisterHandler("/unary", func(path string, header http.Header, recvChan chan *bytes.Buffer, sendChan chan *bytes.Buffer, ctrlCh chan http.Header, errCh chan interface{}) {
 		fmt.Println("path = ", path)
 		fmt.Println("header = ", header)
 
-		rspHeader := make(map[string]string)
-		rspHeader["content-type"] = "application/grpc+proto"
-		rspHeader["grpc-status"] = "Trailer"
-		rspHeader["grpc-message"] = "Trailer"
-		rspHeader["tri-bin"] = "Trailer"
+		rspHeader := make(map[string][]string)
+		rspHeader["content-type"] = []string{"application/grpc+proto"}
+		rspHeader["grpc-status"] = []string{"Trailer"}
+		rspHeader["grpc-message"] = []string{"Trailer"}
+		rspHeader["tri-bin"] = []string{"Trailer"}
 		ctrlCh <- rspHeader
 
 		body := <-recvChan
@@ -30,22 +34,22 @@ func main() {
 		close(sendChan)
 		time.Sleep(time.Second)
 
-		rspHeader2 := make(map[string]string)
-		rspHeader2["grpc-status"] = "grpc-status-val"
-		rspHeader2["grpc-message"] = "grpc-message-val"
-		rspHeader2["tri-bin"] = "tri-bin-val"
+		rspHeader2 := make(map[string][]string)
+		rspHeader2["grpc-status"] = []string{"grpc-status-val"}
+		rspHeader2["grpc-message"] = []string{"grpc-message-val"}
+		rspHeader2["tri-bin"] = []string{"tri-bin-val"}
 		ctrlCh <- rspHeader2
 	})
 
-	svr.RegisterHandler("/stream", func(path string, header map[string]string, recvChan chan *bytes.Buffer, sendChan chan *bytes.Buffer, ctrlCh chan map[string]string) {
+	svr.RegisterHandler("/stream", func(path string, header http.Header, recvChan chan *bytes.Buffer, sendChan chan *bytes.Buffer, ctrlCh chan http.Header, errCh chan interface{}) {
 		fmt.Println("path = ", path)
 		fmt.Println("header = ", header)
 
-		rspHeader := make(map[string]string)
-		rspHeader["content-type"] = "application/grpc+proto"
-		rspHeader["grpc-status"] = "Trailer"
-		rspHeader["grpc-message"] = "Trailer"
-		rspHeader["tri-bin"] = "Trailer"
+		rspHeader := make(map[string][]string)
+		rspHeader["content-type"] = []string{"application/grpc+proto"}
+		rspHeader["grpc-status"] = []string{"Trailer"}
+		rspHeader["grpc-message"] = []string{"Trailer"}
+		rspHeader["tri-bin"] = []string{"Trailer"}
 		ctrlCh <- rspHeader
 
 		body := bytes.NewBuffer([]byte("hello"))
@@ -57,10 +61,10 @@ func main() {
 		close(sendChan)
 		time.Sleep(time.Second)
 
-		rspHeader2 := make(map[string]string)
-		rspHeader2["grpc-status"] = "grpc-status-val"
-		rspHeader2["grpc-message"] = "grpc-message-val"
-		rspHeader2["tri-bin"] = "tri-bin-val"
+		rspHeader2 := make(map[string][]string)
+		rspHeader2["grpc-status"] = []string{"grpc-status-val"}
+		rspHeader2["grpc-message"] = []string{"grpc-message-val"}
+		rspHeader2["tri-bin"] = []string{"tri-bin-val"}
 		ctrlCh <- rspHeader2
 	})
 	svr.Start()
